made the nessary adjusments needed for ep card
