+++ frontend/components/cards/EpisodeCard.tsx
@@ -1,134 +1,79 @@
-import React from "react";
-import {
-  Box,
-  Flex,
-  IconButton,
-  useBreakpointValue,
-  Text,
-  Image,
-  Icon,
-  useToast,
-} from "@chakra-ui/react";
-import { FaPlay } from "react-icons/fa";
-import { BsExplicitFill } from "react-icons/bs";
-import { usePlayer } from "../../utilities/PlayerContext";
-import LikeComponent from "../interactionHub/Likes";
-import { convertTime } from "../../utilities/commonUtils";
-import EpisodeMenu from "./EpisodeMenu";
-import CommentButton from "../interactionHub/buttons/CommentButton";
-import Tipjar from "../interactionHub/Tipjar";
+import { HStack, VStack, Image, Text, IconButton, Icon } from '@chakra-ui/react'
+import { Episode } from '../../types/Interfaces'
+import { usePlayer } from '../../utilities/PlayerContext'
+import { convertTime } from '../../utilities/commonUtils'
+import { Dots, Time, Plays, Play } from '../../public/icons'
+import Likes from '../interactionHub/Likes'
+import CommentButton from '../interactionHub/buttons/CommentButton'
 
+interface EpisodeCardProps {
+  episode: Episode
+  showLike?: boolean
+  showComment?: boolean
+  showMore?: boolean
+}
 
-// Component to display an episode
-const EpisodeCard = ({ episode, inPlaylist, playlistId, inWallet }) => {
-  const { dispatch } = usePlayer();
-  const toast = useToast();
+const EpisodeCard: React.FC<EpisodeCardProps> = ({ episode, showLike = true, showComment = true, showMore = true }) => {
+  const { thumbnailUrl, episodeName, podcastName, duration, likes, playCount, id } = episode
+  const { dispatch } = usePlayer()
 
-  // Handle click on episode
   const handleEpisodeClick = () => {
-    dispatch({
-      type: "PLAY_NOW_QUEUE",
-      payload: episode,
-    });
-  };
-
-  const isMobile = useBreakpointValue({ base: true, md: false });
+    dispatch({ type: 'PLAY_NOW_QUEUE', payload: episode })
+  }
 
   return (
-    <Flex
-      paddingTop={3}
-      paddingBottom={3}
-      mt={3}
-      width="100%"
+    <HStack
+      
+     
+      overflow="hidden"
+      p={'15px'}
+      alignItems="center"
       borderRadius="15px"
-      bg={"rgba(0, 0, 0, 0.2)"}
-      boxShadow="sm"
-      style={{ cursor: "pointer" }}
-      onClick={isMobile ? handleEpisodeClick : null}
-      onDoubleClick={handleEpisodeClick}
+      bg={'az.darkerGrey'}
+      onClick={handleEpisodeClick}
+      _hover={{
+        transform: 'scale(1.01)',
+      }}
+      transition="transform 1s ease-in-out"
+      spacing={3}
     >
-      <Box position="relative" mr={5} onClick={() => handleEpisodeClick()}>
-        <Image
-          boxSize={isMobile ? "0px" : "125px"}
-          src={episode.thumbnailUrl}
-          borderRadius="10%"
-          marginLeft={isMobile ? "0px" : "20px"}
-          mt={1}
-        />
-        {!isMobile && (
-          <IconButton
-            aria-label="Play"
-            icon={<FaPlay />}
-            position="absolute"
-            left="60%"
-            top="50%"
-            transform="translate(-50%, -50%)"
-            variant="ghost"
-            fontSize="25px"
-            shadow={"md"}
-            _hover={{ boxShadow: "lg" }}
-          />
-        )}
-      </Box>
-      <Flex direction="column" flex={1}>
-        {/* Episode Name */}
-        <Text fontWeight="medium" fontSize={isMobile ? "sm" : "2xl"}>
-          {episode.episodeName}
-          {episode.isExplicit && (
-            <Icon
-              as={BsExplicitFill}
-              boxSize={isMobile ? "10px" : "16px"}
-              ml={4}
-            />
-          )}
-          <Text fontSize={isMobile ? "md" : "md"}>🎧 {episode.playCount}</Text>
-        </Text>
-        {/* Episode Details */}
-        <Flex direction="column" fontSize="sm" color={"gray.500"}>
-          {isMobile ? null : <Text>{episode.description}</Text>}
+      <Image src={thumbnailUrl} alt={episodeName} objectFit="cover" width="80px" height="80px" borderRadius="10px" />
 
-          <Text fontWeight="bold" fontSize={isMobile ? "12px" : "md"}>
-            Duration: {convertTime(episode.duration)}
+      <VStack spacing={1} w="310px" flexDir="column" justify="center" align="flex-start" flexShrink={0}>
+        <VStack align="start" spacing={-1}>
+          <Text fontSize="md" fontWeight="bold" color="az.white" noOfLines={2} data-cy={`ticket-episode-${episode.episodeName}`}>
+            {episodeName}
           </Text>
-        </Flex>
-      </Flex>
+          <Text fontSize="xs" color="az.greyish" noOfLines={1}>
+            {podcastName}
+          </Text>
+        </VStack>
+        <HStack>
+          <HStack spacing={1} mr={5}>
+            <Icon as={Plays} color="az.greyish" boxSize={3} />
+            <Text color="az.greyish" fontSize="xs">
+              {playCount}
+            </Text>
+          </HStack>
+          <HStack spacing={1}>
+            <Icon as={Time} color="az.greyish" boxSize={3} />
+            <Text color="az.greyish" fontSize="xs">
+              {convertTime(duration)}
+            </Text>
+          </HStack>
+        </HStack>
+      </VStack>
 
-      {/* Edit and Delete Buttons */}
-      <Flex alignItems="flex-start" style={{ marginRight: "15px" }}>
+      <HStack spacing={1}>
+        <IconButton aria-label={'Play'} icon={<Play />} variant="circle" background="az.red" minWidth="2.5em" size="md" onClick={handleEpisodeClick} />
+        <HStack spacing={0}>
+          {showLike ? <Likes episodeOrCommentId={id} initialLikes={likes} showCount={false} /> : null}
+          {showComment ? <CommentButton episodeId={episode.id} initialComments={0} showCount={false} /> : null}
+          {showMore ? <IconButton aria-label="more" icon={<Icon as={Dots} />} variant="minimal" color="az.greyish" /> : null}
+        </HStack>
+      </HStack>
+    </HStack>
+  )
+}
 
-        {/* Make sure Point is only rendered when inWallet is true */}
-        {inWallet != null && inWallet == true ? (
-          <Tipjar episodeId={episode.id} totalPoint={episode.totalPoints} />
-        ) : (
-          <>
-
-        <CommentButton
-          episodeId={episode.id}
-          initialComments={episode.comments.length}
-          showCount={true}
-        />
-        <Box
-          marginTop="4px"
-          marginLeft="4px"
-          data-cy={`likes-on-${episode.episodeName}-${episode.likes}`}
-        >
-          <LikeComponent
-            episodeOrCommentId={episode.id}
-            initialLikes={episode.likes}
-            showCount={true}
-          />
-        </Box>
-        <EpisodeMenu
-          episode={episode}
-          inPlaylist={inPlaylist}
-          playlistId={playlistId}
-        />
-                </>
-              )}
-
-      </Flex>
-    </Flex>
-  );
-};
-
-export default EpisodeCard;
+export default EpisodeCard
