+++ frontend/pages/Playlist/[playlistId].tsx
@@ -11,7 +11,7 @@ import {
 import type { Episode } from "../../types/Interfaces";
 import type { Playlist } from "../../types/Interfaces";
 import PlaylistHelper from "../../helpers/PlaylistHelper";
-import EpisodeCard from "../../components/cards/EpisodeCard";
+import EpisodeCard from "../../components/playlist/EpisodeCardForPlaylist";
 import { RiPlayList2Fill } from "react-icons/ri";
 import { formatSecToDurationString } from "../../utilities/commonUtils";
 import { ImShuffle } from "react-icons/im";
